requirements.txt
APScheduler==3.10.1
certifi==2024.7.4
greenlet==3.0.3
psutil==5.9.5
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2024.1
sentry-sdk==1.28.1
setuptools==70.2.0
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
